refactor rendering options schema and theme Enum
