+++ libs/newsletters-data-client/src/lib/schemas/themeEnumSchema.ts
@@ -1,0 +1,12 @@
+import { z } from 'zod';
+
+export const themeEnumSchema = z.enum([
+	'news',
+	'opinion',
+	'culture',
+	'sport',
+	'lifestyle',
+	'features',
+]);
+
+export type Theme = z.infer<typeof themeEnumSchema>;
